OLD-requirements.txt
# Core Deep Learning
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# PyTorch Geometric
torch-geometric==2.4.0
torch-scatter
torch-sparse
torch-cluster

# Transformers and NLP
transformers==4.35.0
sentence-transformers==2.2.2
datasets==2.14.0

# Biomedical Data Tools
biopython==1.81
pandas==2.1.0
numpy==1.24.0
networkx==3.1
scipy==1.11.0

# Machine Learning Utilities
scikit-learn==1.3.0
matplotlib==3.8.0
seaborn==0.12.0

# Progress and Logging
tqdm==4.66.0
wandb==0.15.0
pyyaml==6.0

# Development and Testing
pytest==7.4.0
jupyter==1.0.0
ipykernel==6.25.0

# Web Requests
requests==2.31.0

# Additional utilities
python-dotenv==1.0.0
